Create messageRegistry for message parsing code
