--- ironfish/src/network/messages/interfaces/networkMessageHeader.ts
@@ -1,11 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-import { NetworkMessageType } from '../networkMessage'
-
-export interface NetworkMessageHeader {
-  body: Buffer
-  type: NetworkMessageType
-  rpcId?: number
-  nonce?: Buffer
-}
