+++ ironfish/src/network/messages/rpcNetworkMessage.ts
@@ -1,8 +1,9 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+import type { NetworkMessageType } from '../types'
 import bufio from 'bufio'
-import { NetworkMessage, NetworkMessageType } from './networkMessage'
+import { NetworkMessage } from './networkMessage'
 
 export type RpcId = number
 export const RPC_TIMEOUT_MILLIS = 30000
@@ -22,6 +23,13 @@ export abstract class RpcNetworkMessage extends NetworkMessage {
     super(type)
     this.direction = direction
     this.rpcId = rpcId ?? (RpcNetworkMessage.id = ++RpcNetworkMessage.id % 0xffff)
+  }
+
+  static deserializeHeader(buffer: Buffer): { rpcId: number; remaining: Buffer } {
+    const br = bufio.read(buffer, true)
+    const rpcId = br.readU16()
+    const remaining = br.readBytes(br.left())
+    return { rpcId, remaining }
   }
 
   serializeWithMetadata(): Buffer {
