+++ ironfish/src/network/peers/connections/webSocketConnection.ts
@@ -5,12 +5,13 @@
 import type { Logger } from '../../../logger'
 import colors from 'colors/safe'
 import { MetricsMonitor } from '../../../metrics'
+import { parseNetworkMessage } from '../../messageRegistry'
+import { displayNetworkMessageType, NetworkMessage } from '../../messages/networkMessage'
 import {
-  displayNetworkMessageType,
-  NetworkMessage,
+  IsomorphicWebSocket,
+  IsomorphicWebSocketErrorEvent,
   NetworkMessageType,
-} from '../../messages/networkMessage'
-import { IsomorphicWebSocket, IsomorphicWebSocketErrorEvent } from '../../types'
+} from '../../types'
 import { Connection, ConnectionDirection, ConnectionType } from './connection'
 import { NetworkError } from './errors'
 
@@ -87,7 +88,7 @@ export class WebSocketConnection extends Connection {
 
       let message
       try {
-        message = this.parseMessage(event.data)
+        message = parseNetworkMessage(event.data)
         const byteCount = event.data.byteLength
         this.metrics?.p2p_InboundTraffic.add(byteCount)
         this.metrics?.p2p_InboundTraffic_WS.add(byteCount)
